refactor: split api-proxy server.test.js into focused modules (#2623)
Split server.test.js (2,946 lines) into focused test modules:

- server.routing.test.js (692 lines) — URL normalization, routing, OpenCode
- server.auth.test.js (217 lines) — header stripping, auth token resolution
- server.proxy.test.js (405 lines) — WebSocket proxy, request proxying
- server.gemini.test.js (68 lines) — Gemini key param stripping

server.test.js retains remaining tests (1,561 lines): httpProbe, validateApiKeys,
fetchJson, model discovery, provider adapters, billing headers.

All 594 tests pass across 13 test suites.

Closes #2623
