+++ containers/api-proxy/server.test.js
@@ -1,1299 +1,20 @@
 /**
- * Tests for api-proxy server.js
+ * Tests for api-proxy server.js — integration, provider server, and general tests.
+ *
+ * Routing tests → server.routing.test.js
+ * Auth tests → server.auth.test.js
+ * Proxy behavior tests → server.proxy.test.js
+ * Gemini tests → server.gemini.test.js
  */
 
 const http = require('http');
 const https = require('https');
-const tls = require('tls');
 const { EventEmitter } = require('events');
-
-// Functions that live in proxy-utils.js
-const { normalizeApiTarget, normalizeBasePath, buildUpstreamPath, shouldStripHeader, stripGeminiKeyParam, composeBodyTransforms } = require('./proxy-utils');
-
-// Provider-specific functions that live in their respective adapter modules
-const { deriveCopilotApiTarget, deriveGitHubApiTarget, deriveGitHubApiBasePath, resolveCopilotAuthToken, stripBearerPrefix, createCopilotAdapter } = require('./providers/copilot');
-const { resolveOpenCodeRoute } = require('./providers/opencode');
 
 // Core proxy functions that remain in server.js
 const { proxyRequest, proxyWebSocket, httpProbe, validateApiKeys, keyValidationResults, resetKeyValidationState, fetchJson, extractModelIds, fetchStartupModels, reflectEndpoints, healthResponse, cachedModels, resetModelCacheState, makeModelBodyTransform, MODEL_ALIASES, buildModelsJson, writeModelsJson, createProviderServer, extractBillingHeaders } = require('./server');
-
-describe('normalizeApiTarget', () => {
-  it('should strip https:// prefix', () => {
-    expect(normalizeApiTarget('https://my-gateway.example.com')).toBe('my-gateway.example.com');
-  });
-
-  it('should strip http:// prefix', () => {
-    expect(normalizeApiTarget('http://my-gateway.example.com')).toBe('my-gateway.example.com');
-  });
-
-  it('should preserve bare hostname', () => {
-    expect(normalizeApiTarget('api.openai.com')).toBe('api.openai.com');
-  });
-
-  it('should normalize a URL with a path to just the hostname', () => {
-    expect(normalizeApiTarget('https://my-gateway.example.com/some-path')).toBe('my-gateway.example.com');
-  });
-
-  it('should trim whitespace', () => {
-    expect(normalizeApiTarget('  https://api.openai.com  ')).toBe('api.openai.com');
-  });
-
-  it('should return undefined for falsy input', () => {
-    expect(normalizeApiTarget(undefined)).toBeUndefined();
-    expect(normalizeApiTarget('')).toBe('');
-  });
-
-  it('should not strip scheme-like substrings in the middle', () => {
-    expect(normalizeApiTarget('api.https.example.com')).toBe('api.https.example.com');
-  });
-
-  it('should discard port from URL', () => {
-    expect(normalizeApiTarget('https://my-gateway.example.com:8443')).toBe('my-gateway.example.com');
-  });
-
-  it('should discard query and fragment from URL', () => {
-    expect(normalizeApiTarget('https://my-gateway.example.com/path?key=val#frag')).toBe('my-gateway.example.com');
-  });
-});
-
-describe('deriveCopilotApiTarget', () => {
-  let originalEnv;
-
-  beforeEach(() => {
-    // Save original env vars
-    originalEnv = {
-      COPILOT_API_TARGET: process.env.COPILOT_API_TARGET,
-      GITHUB_SERVER_URL: process.env.GITHUB_SERVER_URL,
-    };
-    // Clear env vars before each test
-    delete process.env.COPILOT_API_TARGET;
-    delete process.env.GITHUB_SERVER_URL;
-  });
-
-  afterEach(() => {
-    // Restore original env vars
-    if (originalEnv.COPILOT_API_TARGET !== undefined) {
-      process.env.COPILOT_API_TARGET = originalEnv.COPILOT_API_TARGET;
-    } else {
-      delete process.env.COPILOT_API_TARGET;
-    }
-    if (originalEnv.GITHUB_SERVER_URL !== undefined) {
-      process.env.GITHUB_SERVER_URL = originalEnv.GITHUB_SERVER_URL;
-    } else {
-      delete process.env.GITHUB_SERVER_URL;
-    }
-  });
-
-  describe('COPILOT_API_TARGET env var (highest priority)', () => {
-    it('should return COPILOT_API_TARGET when explicitly set', () => {
-      process.env.COPILOT_API_TARGET = 'custom.api.com';
-      expect(deriveCopilotApiTarget()).toBe('custom.api.com');
-    });
-
-    it('should prefer COPILOT_API_TARGET over GITHUB_SERVER_URL', () => {
-      process.env.COPILOT_API_TARGET = 'custom.api.com';
-      process.env.GITHUB_SERVER_URL = 'https://mycompany.ghe.com';
-      expect(deriveCopilotApiTarget()).toBe('custom.api.com');
-    });
-  });
-
-  describe('GitHub Enterprise Cloud (*.ghe.com)', () => {
-    it('should derive copilot-api.<subdomain>.ghe.com for GHEC tenants', () => {
-      process.env.GITHUB_SERVER_URL = 'https://mycompany.ghe.com';
-      expect(deriveCopilotApiTarget()).toBe('copilot-api.mycompany.ghe.com');
-    });
-
-    it('should handle GHEC URLs with trailing slash', () => {
-      process.env.GITHUB_SERVER_URL = 'https://example.ghe.com/';
-      expect(deriveCopilotApiTarget()).toBe('copilot-api.example.ghe.com');
-    });
-
-    it('should handle GHEC URLs with path components', () => {
-      process.env.GITHUB_SERVER_URL = 'https://acme.ghe.com/some/path';
-      expect(deriveCopilotApiTarget()).toBe('copilot-api.acme.ghe.com');
-    });
-
-    it('should handle multi-part subdomain for GHEC', () => {
-      process.env.GITHUB_SERVER_URL = 'https://dev.mycompany.ghe.com';
-      expect(deriveCopilotApiTarget()).toBe('copilot-api.dev.mycompany.ghe.com');
-    });
-  });
-
-  describe('GitHub Enterprise Server (GHES)', () => {
-    it('should return api.enterprise.githubcopilot.com for GHES', () => {
-      process.env.GITHUB_SERVER_URL = 'https://github.example.com';
-      expect(deriveCopilotApiTarget()).toBe('api.enterprise.githubcopilot.com');
-    });
-
-    it('should handle GHES with IP address', () => {
-      process.env.GITHUB_SERVER_URL = 'https://192.168.1.100';
-      expect(deriveCopilotApiTarget()).toBe('api.enterprise.githubcopilot.com');
-    });
-
-    it('should handle GHES with custom port', () => {
-      process.env.GITHUB_SERVER_URL = 'https://github.internal:8443';
-      expect(deriveCopilotApiTarget()).toBe('api.enterprise.githubcopilot.com');
-    });
-  });
-
-  describe('GitHub.com (public)', () => {
-    it('should return api.githubcopilot.com for github.com', () => {
-      process.env.GITHUB_SERVER_URL = 'https://github.com';
-      expect(deriveCopilotApiTarget()).toBe('api.githubcopilot.com');
-    });
-
-    it('should handle github.com with trailing slash', () => {
-      process.env.GITHUB_SERVER_URL = 'https://github.com/';
-      expect(deriveCopilotApiTarget()).toBe('api.githubcopilot.com');
-    });
-
-    it('should handle github.com with path', () => {
-      process.env.GITHUB_SERVER_URL = 'https://github.com/github/hub';
-      expect(deriveCopilotApiTarget()).toBe('api.githubcopilot.com');
-    });
-  });
-
-  describe('Default behavior', () => {
-    it('should return api.githubcopilot.com when no env vars are set', () => {
-      expect(deriveCopilotApiTarget()).toBe('api.githubcopilot.com');
-    });
-
-    it('should return default when GITHUB_SERVER_URL is empty string', () => {
-      process.env.GITHUB_SERVER_URL = '';
-      expect(deriveCopilotApiTarget()).toBe('api.githubcopilot.com');
-    });
-
-    it('should return default when GITHUB_SERVER_URL is invalid', () => {
-      process.env.GITHUB_SERVER_URL = 'not-a-valid-url';
-      expect(deriveCopilotApiTarget()).toBe('api.githubcopilot.com');
-    });
-
-    it('should return default when GITHUB_SERVER_URL is malformed', () => {
-      process.env.GITHUB_SERVER_URL = 'ht!tp://bad-url';
-      expect(deriveCopilotApiTarget()).toBe('api.githubcopilot.com');
-    });
-  });
-});
-
-describe('deriveGitHubApiTarget', () => {
-  let originalEnv;
-
-  beforeEach(() => {
-    originalEnv = {
-      GITHUB_API_URL: process.env.GITHUB_API_URL,
-      GITHUB_SERVER_URL: process.env.GITHUB_SERVER_URL,
-    };
-    delete process.env.GITHUB_API_URL;
-    delete process.env.GITHUB_SERVER_URL;
-  });
-
-  afterEach(() => {
-    if (originalEnv.GITHUB_API_URL !== undefined) {
-      process.env.GITHUB_API_URL = originalEnv.GITHUB_API_URL;
-    } else {
-      delete process.env.GITHUB_API_URL;
-    }
-    if (originalEnv.GITHUB_SERVER_URL !== undefined) {
-      process.env.GITHUB_SERVER_URL = originalEnv.GITHUB_SERVER_URL;
-    } else {
-      delete process.env.GITHUB_SERVER_URL;
-    }
-  });
-
-  describe('GITHUB_API_URL env var (highest priority)', () => {
-    it('should return hostname from GITHUB_API_URL full URL', () => {
-      process.env.GITHUB_API_URL = 'https://api.github.com';
-      expect(deriveGitHubApiTarget()).toBe('api.github.com');
-    });
-
-    it('should return hostname from GITHUB_API_URL for GHES', () => {
-      process.env.GITHUB_API_URL = 'https://github.internal/api/v3';
-      expect(deriveGitHubApiTarget()).toBe('github.internal');
-    });
-
-    it('should prefer GITHUB_API_URL over GITHUB_SERVER_URL', () => {
-      process.env.GITHUB_API_URL = 'https://api.mycompany.ghe.com';
-      process.env.GITHUB_SERVER_URL = 'https://mycompany.ghe.com';
-      expect(deriveGitHubApiTarget()).toBe('api.mycompany.ghe.com');
-    });
-  });
-
-  describe('GHEC (*.ghe.com)', () => {
-    it('should return api.<subdomain>.ghe.com for GHEC tenant', () => {
-      process.env.GITHUB_SERVER_URL = 'https://mycompany.ghe.com';
-      expect(deriveGitHubApiTarget()).toBe('api.mycompany.ghe.com');
-    });
-
-    it('should handle multiple-level subdomains', () => {
-      process.env.GITHUB_SERVER_URL = 'https://sub.example.ghe.com';
-      expect(deriveGitHubApiTarget()).toBe('api.sub.example.ghe.com');
-    });
-  });
-
-  describe('Default behavior', () => {
-    it('should return api.github.com when no env vars are set', () => {
-      expect(deriveGitHubApiTarget()).toBe('api.github.com');
-    });
-
-    it('should return api.github.com for github.com GITHUB_SERVER_URL', () => {
-      process.env.GITHUB_SERVER_URL = 'https://github.com';
-      expect(deriveGitHubApiTarget()).toBe('api.github.com');
-    });
-
-    it('should return api.github.com for GHES without GITHUB_API_URL', () => {
-      // GHES without an explicit GITHUB_API_URL falls back to api.github.com.
-      // This is a known limitation: GHES deployments should set GITHUB_API_URL explicitly
-      // so deriveGitHubApiTarget() resolves to the correct enterprise API hostname.
-      process.env.GITHUB_SERVER_URL = 'https://github.internal';
-      expect(deriveGitHubApiTarget()).toBe('api.github.com');
-    });
-
-    it('should return api.github.com when GITHUB_SERVER_URL is invalid', () => {
-      process.env.GITHUB_SERVER_URL = 'not-a-valid-url';
-      expect(deriveGitHubApiTarget()).toBe('api.github.com');
-    });
-  });
-});
-
-describe('deriveGitHubApiBasePath', () => {
-  const savedEnv = {};
-
-  beforeEach(() => {
-    savedEnv.GITHUB_API_URL = process.env.GITHUB_API_URL;
-    delete process.env.GITHUB_API_URL;
-  });
-
-  afterEach(() => {
-    if (savedEnv.GITHUB_API_URL !== undefined) {
-      process.env.GITHUB_API_URL = savedEnv.GITHUB_API_URL;
-    } else {
-      delete process.env.GITHUB_API_URL;
-    }
-  });
-
-  it('should return empty string when GITHUB_API_URL is not set', () => {
-    expect(deriveGitHubApiBasePath()).toBe('');
-  });
-
-  it('should extract /api/v3 from GHES-style GITHUB_API_URL', () => {
-    process.env.GITHUB_API_URL = 'https://ghes.example.com/api/v3';
-    expect(deriveGitHubApiBasePath()).toBe('/api/v3');
-  });
-
-  it('should return empty string for github.com API URL (no path)', () => {
-    process.env.GITHUB_API_URL = 'https://api.github.com';
-    expect(deriveGitHubApiBasePath()).toBe('');
-  });
-
-  it('should strip trailing slashes', () => {
-    process.env.GITHUB_API_URL = 'https://ghes.example.com/api/v3/';
-    expect(deriveGitHubApiBasePath()).toBe('/api/v3');
-  });
-
-  it('should return empty string for invalid URL', () => {
-    process.env.GITHUB_API_URL = '://invalid';
-    expect(deriveGitHubApiBasePath()).toBe('');
-  });
-});
-
-describe('normalizeBasePath', () => {
-  it('should return empty string for undefined', () => {
-    expect(normalizeBasePath(undefined)).toBe('');
-  });
-
-  it('should return empty string for null', () => {
-    expect(normalizeBasePath(null)).toBe('');
-  });
-
-  it('should return empty string for empty string', () => {
-    expect(normalizeBasePath('')).toBe('');
-  });
-
-  it('should return empty string for whitespace-only string', () => {
-    expect(normalizeBasePath('   ')).toBe('');
-  });
-
-  it('should preserve a well-formed path', () => {
-    expect(normalizeBasePath('/serving-endpoints')).toBe('/serving-endpoints');
-  });
-
-  it('should add leading slash when missing', () => {
-    expect(normalizeBasePath('serving-endpoints')).toBe('/serving-endpoints');
-  });
-
-  it('should strip trailing slash', () => {
-    expect(normalizeBasePath('/serving-endpoints/')).toBe('/serving-endpoints');
-  });
-
-  it('should handle multi-segment paths', () => {
-    expect(normalizeBasePath('/openai/deployments/gpt-4')).toBe('/openai/deployments/gpt-4');
-  });
-
-  it('should normalize a path missing the leading slash and with trailing slash', () => {
-    expect(normalizeBasePath('openai/deployments/gpt-4/')).toBe('/openai/deployments/gpt-4');
-  });
-
-  it('should preserve a root-only path', () => {
-    expect(normalizeBasePath('/')).toBe('/');
-  });
-});
-
-describe('buildUpstreamPath', () => {
-  const HOST = 'api.example.com';
-
-  describe('no base path (empty string)', () => {
-    it('should return the request path unchanged when basePath is empty', () => {
-      expect(buildUpstreamPath('/v1/chat/completions', HOST, '')).toBe('/v1/chat/completions');
-    });
-
-    it('should preserve query string when basePath is empty', () => {
-      expect(buildUpstreamPath('/v1/chat/completions?stream=true', HOST, '')).toBe('/v1/chat/completions?stream=true');
-    });
-
-    it('should preserve multiple query params when basePath is empty', () => {
-      expect(buildUpstreamPath('/v1/models?limit=10&order=asc', HOST, '')).toBe('/v1/models?limit=10&order=asc');
-    });
-
-    it('should handle root path with no base path', () => {
-      expect(buildUpstreamPath('/', HOST, '')).toBe('/');
-    });
-
-    it('should reject protocol-relative URLs to prevent host override', () => {
-      expect(() => buildUpstreamPath('//evil.com/v1/chat/completions', HOST, ''))
-        .toThrow('URL must be a relative origin-form path');
-    });
-  });
-
-  describe('Databricks serving-endpoints (single-segment base path)', () => {
-    it('should prepend /serving-endpoints to chat completions path', () => {
-      expect(buildUpstreamPath('/v1/chat/completions', HOST, '/serving-endpoints'))
-        .toBe('/serving-endpoints/v1/chat/completions');
-    });
-
-    it('should prepend /serving-endpoints and preserve query string', () => {
-      expect(buildUpstreamPath('/v1/chat/completions?stream=true', HOST, '/serving-endpoints'))
-        .toBe('/serving-endpoints/v1/chat/completions?stream=true');
-    });
-
-    it('should prepend /serving-endpoints to models path', () => {
-      expect(buildUpstreamPath('/v1/models', HOST, '/serving-endpoints'))
-        .toBe('/serving-endpoints/v1/models');
-    });
-
-    it('should prepend /serving-endpoints to embeddings path', () => {
-      expect(buildUpstreamPath('/v1/embeddings', HOST, '/serving-endpoints'))
-        .toBe('/serving-endpoints/v1/embeddings');
-    });
-  });
-
-  describe('Azure OpenAI deployments (multi-segment base path)', () => {
-    it('should prepend Azure deployment path to chat completions', () => {
-      expect(buildUpstreamPath('/chat/completions', HOST, '/openai/deployments/gpt-4'))
-        .toBe('/openai/deployments/gpt-4/chat/completions');
-    });
-
-    it('should prepend Azure deployment path and preserve api-version query param', () => {
-      expect(buildUpstreamPath('/chat/completions?api-version=2024-02-01', HOST, '/openai/deployments/gpt-4'))
-        .toBe('/openai/deployments/gpt-4/chat/completions?api-version=2024-02-01');
-    });
-
-    it('should handle a deeply nested Azure deployment name', () => {
-      expect(buildUpstreamPath('/chat/completions', HOST, '/openai/deployments/my-custom-gpt-4-deployment'))
-        .toBe('/openai/deployments/my-custom-gpt-4-deployment/chat/completions');
-    });
-  });
-
-  describe('Anthropic custom target with base path', () => {
-    it('should prepend /anthropic to messages endpoint', () => {
-      expect(buildUpstreamPath('/v1/messages', 'proxy.corporate.com', '/anthropic'))
-        .toBe('/anthropic/v1/messages');
-    });
-
-    it('should preserve Anthropic query params', () => {
-      expect(buildUpstreamPath('/v1/messages?beta=true', 'proxy.corporate.com', '/anthropic'))
-        .toBe('/anthropic/v1/messages?beta=true');
-    });
-  });
-
-  describe('path preservation for real-world API endpoints', () => {
-    it('should preserve /v1/chat/completions exactly (OpenAI standard path)', () => {
-      expect(buildUpstreamPath('/v1/chat/completions', 'api.openai.com', ''))
-        .toBe('/v1/chat/completions');
-    });
-
-    it('should map unversioned /responses to /v1/responses when basePath is /v1 (OpenAI default)', () => {
-      // The OpenAI adapter passes basePath='/v1' for the public endpoint.
-      // buildUpstreamPath is now provider-agnostic; the /v1 prefix comes from the adapter.
-      expect(buildUpstreamPath('/responses', 'api.openai.com', '/v1'))
-        .toBe('/v1/responses');
-    });
-
-    it('should preserve already-versioned OpenAI responses path with /v1 basePath', () => {
-      expect(buildUpstreamPath('/v1/responses', 'api.openai.com', '/v1'))
-        .toBe('/v1/responses');
-    });
-
-    it('should map unversioned /responses to /v1/responses when basePath is /v1 (host-with-port variant)', () => {
-      // basePath='/v1' is the canonical form; the OpenAI adapter normalises the target.
-      expect(buildUpstreamPath('/responses', 'api.openai.com', '/v1'))
-        .toBe('/v1/responses');
-    });
-
-    it('should preserve /v1/messages exactly (Anthropic standard path)', () => {
-      expect(buildUpstreamPath('/v1/messages', 'api.anthropic.com', ''))
-        .toBe('/v1/messages');
-    });
-
-    it('should handle URL-encoded characters in path', () => {
-      // %2F is preserved by the URL parser (an encoded slash stays encoded)
-      expect(buildUpstreamPath('/v1/models/gpt-4%2Fturbo', HOST, '/serving-endpoints'))
-        .toBe('/serving-endpoints/v1/models/gpt-4%2Fturbo');
-    });
-
-    it('should handle hash fragment being ignored (not forwarded in HTTP requests)', () => {
-      // Hash fragments are never sent to the server; URL parser drops them
-      expect(buildUpstreamPath('/v1/chat/completions#fragment', HOST, '/serving-endpoints'))
-        .toBe('/serving-endpoints/v1/chat/completions');
-    });
-
-    it('should drop empty query string marker', () => {
-      expect(buildUpstreamPath('/v1/chat/completions?', HOST, '/serving-endpoints'))
-        .toBe('/serving-endpoints/v1/chat/completions');
-    });
-  });
-
-  describe('with normalized API target (gh-aw#25137 regression)', () => {
-    it('should produce correct path when target was already normalized', () => {
-      // normalizeApiTarget('https://my-gateway.example.com/some-path')
-      // returns 'my-gateway.example.com' (hostname only)
-      const target = 'my-gateway.example.com';
-      expect(buildUpstreamPath('/v1/messages', target, ''))
-        .toBe('/v1/messages');
-    });
-
-    it('should not force /v1 for non-OpenAI custom targets', () => {
-      const target = 'my-gateway.example.com';
-      expect(buildUpstreamPath('/responses', target, ''))
-        .toBe('/responses');
-    });
-
-    it('should produce wrong hostname if scheme is NOT stripped (demonstrating the bug)', () => {
-      // Without normalizeApiTarget, the scheme-prefixed value causes
-      // new URL() to parse 'https' as the hostname instead of the real host
-      const badTarget = 'https://my-gateway.example.com';
-      const targetUrl = new URL('/v1/messages', `https://${badTarget}`);
-      // Node parses this as hostname='https', not 'my-gateway.example.com'
-      expect(targetUrl.hostname).not.toBe('my-gateway.example.com');
-    });
-  });
-});
-
-describe('shouldStripHeader', () => {
-  it('should strip authorization header', () => {
-    expect(shouldStripHeader('authorization')).toBe(true);
-    expect(shouldStripHeader('Authorization')).toBe(true);
-  });
-
-  it('should strip x-api-key header', () => {
-    expect(shouldStripHeader('x-api-key')).toBe(true);
-    expect(shouldStripHeader('X-Api-Key')).toBe(true);
-  });
-
-  it('should strip x-goog-api-key header (Gemini placeholder must be stripped)', () => {
-    expect(shouldStripHeader('x-goog-api-key')).toBe(true);
-    expect(shouldStripHeader('X-Goog-Api-Key')).toBe(true);
-  });
-
-  it('should strip proxy-authorization header', () => {
-    expect(shouldStripHeader('proxy-authorization')).toBe(true);
-  });
-
-  it('should strip x-forwarded-* headers', () => {
-    expect(shouldStripHeader('x-forwarded-for')).toBe(true);
-    expect(shouldStripHeader('x-forwarded-host')).toBe(true);
-  });
-
-  it('should not strip content-type header', () => {
-    expect(shouldStripHeader('content-type')).toBe(false);
-  });
-
-  it('should not strip anthropic-version header', () => {
-    expect(shouldStripHeader('anthropic-version')).toBe(false);
-  });
-});
-
-describe('stripGeminiKeyParam', () => {
-  it('should remove the key= query parameter', () => {
-    expect(stripGeminiKeyParam('/v1/models/gemini-pro:generateContent?key=placeholder'))
-      .toBe('/v1/models/gemini-pro:generateContent');
-  });
-
-  it('should remove key= while preserving other query parameters', () => {
-    expect(stripGeminiKeyParam('/v1/models/gemini-pro:generateContent?key=placeholder&alt=json'))
-      .toBe('/v1/models/gemini-pro:generateContent?alt=json');
-  });
-
-  it('should return path unchanged when no key= parameter is present', () => {
-    expect(stripGeminiKeyParam('/v1/models/gemini-pro:generateContent'))
-      .toBe('/v1/models/gemini-pro:generateContent');
-  });
-
-  it('should return path unchanged when only unrelated query parameters exist', () => {
-    expect(stripGeminiKeyParam('/v1/models/gemini-pro:generateContent?alt=json&stream=true'))
-      .toBe('/v1/models/gemini-pro:generateContent?alt=json&stream=true');
-  });
-
-  it('should handle root path without key param', () => {
-    expect(stripGeminiKeyParam('/')).toBe('/');
-  });
-
-  it('should handle path with only key= param, leaving no trailing ?', () => {
-    // URL.search returns '' when no params remain after deletion
-    const result = stripGeminiKeyParam('/v1/generateContent?key=abc');
-    expect(result).toBe('/v1/generateContent');
-  });
-
-  it('should remove the apiKey= query parameter', () => {
-    expect(stripGeminiKeyParam('/v1/models/gemini-pro:generateContent?apiKey=placeholder'))
-      .toBe('/v1/models/gemini-pro:generateContent');
-  });
-
-  it('should remove the api_key= query parameter', () => {
-    expect(stripGeminiKeyParam('/v1/models/gemini-pro:generateContent?api_key=placeholder'))
-      .toBe('/v1/models/gemini-pro:generateContent');
-  });
-
-  it('should remove apiKey= while preserving other query parameters', () => {
-    expect(stripGeminiKeyParam('/v1/models/gemini-pro:generateContent?apiKey=placeholder&alt=json'))
-      .toBe('/v1/models/gemini-pro:generateContent?alt=json');
-  });
-
-  it('should remove api_key= while preserving other query parameters', () => {
-    expect(stripGeminiKeyParam('/v1/models/gemini-pro:generateContent?api_key=placeholder&alt=json'))
-      .toBe('/v1/models/gemini-pro:generateContent?alt=json');
-  });
-
-  it('should remove all auth params when multiple variants are present', () => {
-    expect(stripGeminiKeyParam('/v1/models/gemini-pro:generateContent?key=foo&apiKey=bar&api_key=baz&alt=json'))
-      .toBe('/v1/models/gemini-pro:generateContent?alt=json');
-  });
-
-  it('should handle path with only api_key= param, leaving no trailing ?', () => {
-    const result = stripGeminiKeyParam('/v1/generateContent?api_key=abc');
-    expect(result).toBe('/v1/generateContent');
-  });
-});
-
-// ── Helpers for proxyWebSocket tests ──────────────────────────────────────────
-
-/** Create a minimal mock socket with write/destroy spies. */
-function makeMockSocket() {
-  const s = new EventEmitter();
-  s.write = jest.fn();
-  s.destroy = jest.fn();
-  s.pipe = jest.fn();
-  s.writable = true;
-  s.destroyed = false;
-  return s;
-}
-
-/** Create a mock HTTP request for a WebSocket upgrade. */
-function makeUpgradeReq(overrides = {}) {
-  return {
-    url: '/v1/responses',
-    headers: {
-      'upgrade': 'websocket',
-      'connection': 'Upgrade',
-      'sec-websocket-key': 'test-ws-key==',
-      'sec-websocket-version': '13',
-      'host': '172.30.0.30',
-      ...overrides.headers,
-    },
-    ...overrides,
-  };
-}
-
-describe('proxyWebSocket', () => {
-  afterEach(() => {
-    jest.restoreAllMocks();
-  });
-
-  // ── Request validation ─────────────────────────────────────────────────────
-
-  describe('request validation', () => {
-    it('rejects a non-WebSocket upgrade (e.g. h2c) with 400', () => {
-      const socket = makeMockSocket();
-      proxyWebSocket(makeUpgradeReq({ headers: { 'upgrade': 'h2c' } }), socket, Buffer.alloc(0), 'api.openai.com', {}, 'openai');
-      expect(socket.write).toHaveBeenCalledWith(expect.stringContaining('HTTP/1.1 400 Bad Request'));
-      expect(socket.destroy).toHaveBeenCalled();
-    });
-
-    it('rejects an upgrade with no Upgrade header with 400', () => {
-      const socket = makeMockSocket();
-      const req = makeUpgradeReq();
-      delete req.headers['upgrade'];
-      proxyWebSocket(req, socket, Buffer.alloc(0), 'api.openai.com', {}, 'openai');
-      expect(socket.write).toHaveBeenCalledWith(expect.stringContaining('HTTP/1.1 400 Bad Request'));
-      expect(socket.destroy).toHaveBeenCalled();
-    });
-
-    it('rejects an absolute URL with 400 (SSRF prevention)', () => {
-      const socket = makeMockSocket();
-      proxyWebSocket(makeUpgradeReq({ url: 'https://evil.com/v1/responses' }), socket, Buffer.alloc(0), 'api.openai.com', {}, 'openai');
-      expect(socket.write).toHaveBeenCalledWith(expect.stringContaining('HTTP/1.1 400 Bad Request'));
-      expect(socket.destroy).toHaveBeenCalled();
-    });
-
-    it('rejects a protocol-relative URL with 400 (SSRF prevention)', () => {
-      const socket = makeMockSocket();
-      proxyWebSocket(makeUpgradeReq({ url: '//evil.com/v1/responses' }), socket, Buffer.alloc(0), 'api.openai.com', {}, 'openai');
-      expect(socket.write).toHaveBeenCalledWith(expect.stringContaining('HTTP/1.1 400 Bad Request'));
-      expect(socket.destroy).toHaveBeenCalled();
-    });
-
-    it('rejects a null URL with 400', () => {
-      const socket = makeMockSocket();
-      proxyWebSocket(makeUpgradeReq({ url: null }), socket, Buffer.alloc(0), 'api.openai.com', {}, 'openai');
-      expect(socket.write).toHaveBeenCalledWith(expect.stringContaining('HTTP/1.1 400 Bad Request'));
-      expect(socket.destroy).toHaveBeenCalled();
-    });
-  });
-
-  // ── Proxy config errors ────────────────────────────────────────────────────
-
-  describe('proxy configuration errors', () => {
-    it('returns 502 when HTTPS_PROXY is not configured', () => {
-      // The module was loaded without HTTPS_PROXY; proxyWebSocket should fail-safe.
-      const socket = makeMockSocket();
-      proxyWebSocket(makeUpgradeReq(), socket, Buffer.alloc(0), 'api.openai.com', {}, 'openai');
-      expect(socket.write).toHaveBeenCalledWith(expect.stringContaining('HTTP/1.1 502 Bad Gateway'));
-      expect(socket.destroy).toHaveBeenCalled();
-    });
-  });
-
-  // ── Network tunnel tests (module loaded with HTTPS_PROXY set) ─────────────
-
-  describe('CONNECT tunnel and auth injection', () => {
-    let wsProxy;
-
-    beforeAll(() => {
-      // Re-require server with HTTPS_PROXY so proxyWebSocket uses the proxy URL.
-      process.env.HTTPS_PROXY = 'http://127.0.0.1:3128';
-      jest.resetModules();
-      wsProxy = require('./server').proxyWebSocket;
-    });
-
-    afterAll(() => {
-      delete process.env.HTTPS_PROXY;
-      jest.resetModules();
-    });
-
-    it('returns 502 when the CONNECT response is not 200', () => {
-      const socket = makeMockSocket();
-      const connectReq = new EventEmitter();
-      connectReq.end = jest.fn();
-      const tunnel = makeMockSocket();
-
-      jest.spyOn(http, 'request').mockReturnValue(connectReq);
-      setImmediate(() => connectReq.emit('connect', { statusCode: 407 }, tunnel));
-
-      wsProxy(makeUpgradeReq(), socket, Buffer.alloc(0), 'api.openai.com', { 'Authorization': 'Bearer key' }, 'openai');
-
-      return new Promise(resolve => setImmediate(() => {
-        expect(socket.write).toHaveBeenCalledWith(expect.stringContaining('HTTP/1.1 502 Bad Gateway'));
-        expect(socket.destroy).toHaveBeenCalled();
-        expect(tunnel.destroy).toHaveBeenCalled();
-        resolve();
-      }));
-    });
-
-    it('returns 502 when the CONNECT request emits an error', () => {
-      const socket = makeMockSocket();
-      const connectReq = new EventEmitter();
-      connectReq.end = jest.fn();
-
-      jest.spyOn(http, 'request').mockReturnValue(connectReq);
-      setImmediate(() => connectReq.emit('error', new Error('connection refused')));
-
-      wsProxy(makeUpgradeReq(), socket, Buffer.alloc(0), 'api.openai.com', { 'Authorization': 'Bearer key' }, 'openai');
-
-      return new Promise(resolve => setImmediate(() => {
-        expect(socket.write).toHaveBeenCalledWith(expect.stringContaining('HTTP/1.1 502 Bad Gateway'));
-        expect(socket.destroy).toHaveBeenCalled();
-        resolve();
-      }));
-    });
-
-    it('returns 502 when TLS handshake fails', () => {
-      const socket = makeMockSocket();
-      const connectReq = new EventEmitter();
-      connectReq.end = jest.fn();
-      const tunnel = makeMockSocket();
-      const tlsSocket = new EventEmitter();
-      tlsSocket.write = jest.fn();
-      tlsSocket.destroy = jest.fn();
-      tlsSocket.pipe = jest.fn();
-
-      jest.spyOn(http, 'request').mockReturnValue(connectReq);
-      jest.spyOn(tls, 'connect').mockReturnValue(tlsSocket);
-
-      setImmediate(() => {
-        connectReq.emit('connect', { statusCode: 200 }, tunnel);
-        setImmediate(() => tlsSocket.emit('error', new Error('certificate unknown')));
-      });
-
-      wsProxy(makeUpgradeReq(), socket, Buffer.alloc(0), 'api.openai.com', { 'Authorization': 'Bearer key' }, 'openai');
-
-      return new Promise(resolve => setTimeout(() => {
-        expect(socket.write).toHaveBeenCalledWith(expect.stringContaining('HTTP/1.1 502 Bad Gateway'));
-        expect(socket.destroy).toHaveBeenCalled();
-        resolve();
-      }, 30));
-    });
-
-    it('injects Authorization header and fixes Host header in the upgrade request', () => {
-      const socket = makeMockSocket();
-      const connectReq = new EventEmitter();
-      connectReq.end = jest.fn();
-      const tunnel = makeMockSocket();
-      const tlsSocket = new EventEmitter();
-      tlsSocket.write = jest.fn();
-      tlsSocket.destroy = jest.fn();
-      tlsSocket.pipe = jest.fn();
-
-      jest.spyOn(http, 'request').mockReturnValue(connectReq);
-      jest.spyOn(tls, 'connect').mockReturnValue(tlsSocket);
-
-      setImmediate(() => {
-        connectReq.emit('connect', { statusCode: 200 }, tunnel);
-        setImmediate(() => tlsSocket.emit('secureConnect'));
-      });
-
-      wsProxy(makeUpgradeReq(), socket, Buffer.alloc(0), 'api.openai.com', { 'Authorization': 'Bearer secret' }, 'openai');
-
-      return new Promise(resolve => setTimeout(() => {
-        // The upgrade request is written as a string to tlsSocket
-        const upgradeWrite = tlsSocket.write.mock.calls.find(
-          c => typeof c[0] === 'string' && c[0].startsWith('GET ')
-        );
-        expect(upgradeWrite).toBeDefined();
-        const upgradeReqStr = upgradeWrite[0];
-        expect(upgradeReqStr).toContain('Authorization: Bearer secret');
-        expect(upgradeReqStr).toContain('host: api.openai.com');
-        // Both sides should be piped
-        expect(tlsSocket.pipe).toHaveBeenCalledWith(socket);
-        expect(socket.pipe).toHaveBeenCalledWith(tlsSocket);
-        resolve();
-      }, 30));
-    });
-
-    it('strips client-supplied auth headers before forwarding', () => {
-      const socket = makeMockSocket();
-      const connectReq = new EventEmitter();
-      connectReq.end = jest.fn();
-      const tunnel = makeMockSocket();
-      const tlsSocket = new EventEmitter();
-      tlsSocket.write = jest.fn();
-      tlsSocket.destroy = jest.fn();
-      tlsSocket.pipe = jest.fn();
-
-      jest.spyOn(http, 'request').mockReturnValue(connectReq);
-      jest.spyOn(tls, 'connect').mockReturnValue(tlsSocket);
-
-      setImmediate(() => {
-        connectReq.emit('connect', { statusCode: 200 }, tunnel);
-        setImmediate(() => tlsSocket.emit('secureConnect'));
-      });
-
-      const req = makeUpgradeReq({
-        headers: {
-          'upgrade': 'websocket',
-          'authorization': 'Bearer client-supplied',  // must be stripped
-          'x-api-key': 'client-api-key',              // must be stripped
-          'sec-websocket-key': 'ws-key==',
-          'sec-websocket-version': '13',
-        },
-      });
-
-      wsProxy(req, socket, Buffer.alloc(0), 'api.openai.com', { 'Authorization': 'Bearer injected' }, 'openai');
-
-      return new Promise(resolve => setTimeout(() => {
-        const upgradeWrite = tlsSocket.write.mock.calls.find(
-          c => typeof c[0] === 'string' && c[0].startsWith('GET ')
-        );
-        expect(upgradeWrite).toBeDefined();
-        const upgradeReqStr = upgradeWrite[0];
-        // Client-supplied auth is stripped; injected auth is present
-        expect(upgradeReqStr).not.toContain('client-supplied');
-        expect(upgradeReqStr).not.toContain('client-api-key');
-        expect(upgradeReqStr).toContain('Bearer injected');
-        resolve();
-      }, 30));
-    });
-
-    it('forwards the CONNECT request to the configured Squid proxy host/port', () => {
-      const socket = makeMockSocket();
-      const connectReq = new EventEmitter();
-      connectReq.end = jest.fn();
-
-      let capturedOptions;
-      jest.spyOn(http, 'request').mockImplementation((options) => {
-        capturedOptions = options;
-        return connectReq;
-      });
-
-      wsProxy(makeUpgradeReq(), socket, Buffer.alloc(0), 'api.openai.com', {}, 'openai');
-
-      expect(capturedOptions).toBeDefined();
-      expect(capturedOptions.method).toBe('CONNECT');
-      expect(capturedOptions.path).toBe('api.openai.com:443');
-      expect(capturedOptions.host).toBe('127.0.0.1');
-      expect(capturedOptions.port).toBe(3128);
-    });
-
-    it('forwards buffered head bytes to the upstream after upgrade', () => {
-      const socket = makeMockSocket();
-      const connectReq = new EventEmitter();
-      connectReq.end = jest.fn();
-      const tunnel = makeMockSocket();
-      const tlsSocket = new EventEmitter();
-      tlsSocket.write = jest.fn();
-      tlsSocket.destroy = jest.fn();
-      tlsSocket.pipe = jest.fn();
-
-      jest.spyOn(http, 'request').mockReturnValue(connectReq);
-      jest.spyOn(tls, 'connect').mockReturnValue(tlsSocket);
-
-      setImmediate(() => {
-        connectReq.emit('connect', { statusCode: 200 }, tunnel);
-        setImmediate(() => tlsSocket.emit('secureConnect'));
-      });
-
-      const headBytes = Buffer.from([0x81, 0x05, 0x48, 0x65, 0x6c, 0x6c, 0x6f]); // WS text frame: FIN=1, opcode=1, len=5, payload='Hello'
-      wsProxy(makeUpgradeReq(), socket, headBytes, 'api.openai.com', { 'Authorization': 'Bearer k' }, 'openai');
-
-      return new Promise(resolve => setTimeout(() => {
-        // The head buffer should have been written to tlsSocket
-        const bufWrite = tlsSocket.write.mock.calls.find(c => Buffer.isBuffer(c[0]));
-        expect(bufWrite).toBeDefined();
-        expect(bufWrite[0]).toEqual(headBytes);
-        resolve();
-      }, 30));
-    });
-  });
-});
-
-describe('stripBearerPrefix', () => {
-  it('strips "Bearer " prefix from a token value', () => {
-    expect(stripBearerPrefix('Bearer sk-or-v1-abc')).toBe('sk-or-v1-abc');
-  });
-
-  it('strips "Bearer " prefix case-insensitively', () => {
-    expect(stripBearerPrefix('bearer sk-or-v1-abc')).toBe('sk-or-v1-abc');
-    expect(stripBearerPrefix('BEARER sk-or-v1-abc')).toBe('sk-or-v1-abc');
-  });
-
-  it('strips leading whitespace before "Bearer "', () => {
-    expect(stripBearerPrefix('  Bearer sk-or-v1-abc')).toBe('sk-or-v1-abc');
-  });
-
-  it('returns value unchanged when no "Bearer " prefix is present', () => {
-    expect(stripBearerPrefix('sk-or-v1-abc')).toBe('sk-or-v1-abc');
-    expect(stripBearerPrefix('gho_abc123')).toBe('gho_abc123');
-  });
-
-  it('does not strip "Bearer" without a following space', () => {
-    expect(stripBearerPrefix('BearerToken123')).toBe('BearerToken123');
-  });
-
-  it('returns undefined when value is only "Bearer " (nothing after prefix)', () => {
-    expect(stripBearerPrefix('Bearer ')).toBeUndefined();
-    expect(stripBearerPrefix('Bearer   ')).toBeUndefined();
-  });
-
-  it('returns undefined for empty or whitespace-only input', () => {
-    expect(stripBearerPrefix('')).toBeUndefined();
-    expect(stripBearerPrefix('   ')).toBeUndefined();
-    expect(stripBearerPrefix(undefined)).toBeUndefined();
-  });
-
-  it('trims surrounding whitespace from the token', () => {
-    expect(stripBearerPrefix('  sk-or-v1-abc  ')).toBe('sk-or-v1-abc');
-  });
-});
-
-describe('resolveCopilotAuthToken', () => {
-  it('should return COPILOT_GITHUB_TOKEN when only it is set', () => {
-    expect(resolveCopilotAuthToken({ COPILOT_GITHUB_TOKEN: 'gho_abc123' })).toBe('gho_abc123');
-  });
-
-  it('should return COPILOT_API_KEY when only it is set', () => {
-    expect(resolveCopilotAuthToken({ COPILOT_API_KEY: 'sk-byok-key' })).toBe('sk-byok-key');
-  });
-
-  it('should prefer COPILOT_GITHUB_TOKEN over COPILOT_API_KEY when both are set', () => {
-    expect(resolveCopilotAuthToken({
-      COPILOT_GITHUB_TOKEN: 'gho_abc123',
-      COPILOT_API_KEY: 'sk-byok-key',
-    })).toBe('gho_abc123');
-  });
-
-  it('should return undefined when neither is set', () => {
-    expect(resolveCopilotAuthToken({})).toBeUndefined();
-  });
-
-  it('should return undefined for empty strings', () => {
-    expect(resolveCopilotAuthToken({ COPILOT_GITHUB_TOKEN: '', COPILOT_API_KEY: '' })).toBeUndefined();
-  });
-
-  it('should return undefined for whitespace-only values', () => {
-    expect(resolveCopilotAuthToken({ COPILOT_GITHUB_TOKEN: '  ', COPILOT_API_KEY: '  \n' })).toBeUndefined();
-  });
-
-  it('should trim whitespace from token values', () => {
-    expect(resolveCopilotAuthToken({ COPILOT_API_KEY: '  sk-byok-key  ' })).toBe('sk-byok-key');
-  });
-
-  it('should fall back to COPILOT_API_KEY when COPILOT_GITHUB_TOKEN is whitespace-only', () => {
-    expect(resolveCopilotAuthToken({
-      COPILOT_GITHUB_TOKEN: '  ',
-      COPILOT_API_KEY: 'sk-byok-key',
-    })).toBe('sk-byok-key');
-  });
-
-  // Integration: verify that Bearer-prefix stripping (via stripBearerPrefix) is
-  // applied to both token sources when resolving.
-
-  it('strips "Bearer " prefix from COPILOT_API_KEY when resolving', () => {
-    expect(resolveCopilotAuthToken({ COPILOT_API_KEY: 'Bearer sk-or-v1-abc' })).toBe('sk-or-v1-abc');
-  });
-
-  it('strips "Bearer " prefix from COPILOT_GITHUB_TOKEN when resolving', () => {
-    expect(resolveCopilotAuthToken({ COPILOT_GITHUB_TOKEN: 'Bearer gho_abc123' })).toBe('gho_abc123');
-  });
-
-  it('prefers stripped COPILOT_GITHUB_TOKEN over stripped COPILOT_API_KEY', () => {
-    expect(resolveCopilotAuthToken({
-      COPILOT_GITHUB_TOKEN: 'Bearer gho_abc123',
-      COPILOT_API_KEY: 'Bearer sk-byok-key',
-    })).toBe('gho_abc123');
-  });
-});
-
-// ── createCopilotAdapter — BYOK auth header format ───────────────────────────
-//
-// These tests guard against the "badly formatted Authorization header" bug in
-// BYOK mode where the sidecar is configured with COPILOT_API_KEY (the real key
-// held by the sidecar) and could produce "Authorization: Bearer Bearer <key>"
-// if the COPILOT_API_KEY value already contained the "Bearer " prefix.
-// They also verify that the header injected for inference requests is exactly
-// "Bearer <key>" and that the Copilot-Integration-Id header is present.
-
-describe('createCopilotAdapter — BYOK getAuthHeaders', () => {
-  const fakeReq = { url: '/v1/chat/completions', method: 'POST', headers: {} };
-  const fakeModelsReq = { url: '/models', method: 'GET', headers: {} };
-
-  it('injects Authorization: Bearer <key> for BYOK inference request', () => {
-    const adapter = createCopilotAdapter({ COPILOT_API_KEY: 'sk-or-v1-abc123' });
-    const headers = adapter.getAuthHeaders(fakeReq);
-    expect(headers['Authorization']).toBe('Bearer sk-or-v1-abc123');
-  });
-
-  it('injects Copilot-Integration-Id header for BYOK inference request', () => {
-    const adapter = createCopilotAdapter({ COPILOT_API_KEY: 'sk-or-v1-abc123' });
-    const headers = adapter.getAuthHeaders(fakeReq);
-    expect(headers['Copilot-Integration-Id']).toBe('copilot-developer-cli');
-  });
-
-  it('prevents double "Bearer " prefix when API key already contains "Bearer " prefix (BYOK bug fix)', () => {
-    const adapter = createCopilotAdapter({ COPILOT_API_KEY: 'Bearer sk-or-v1-abc123' });
-    const headers = adapter.getAuthHeaders(fakeReq);
-    // Must NOT be "Bearer Bearer sk-or-v1-abc123"
-    expect(headers['Authorization']).toBe('Bearer sk-or-v1-abc123');
-    expect(headers['Authorization']).not.toContain('Bearer Bearer');
-  });
-
-  it('strips "Bearer " prefix case-insensitively from API key', () => {
-    const adapter = createCopilotAdapter({ COPILOT_API_KEY: 'BEARER sk-or-v1-abc123' });
-    const headers = adapter.getAuthHeaders(fakeReq);
-    expect(headers['Authorization']).toBe('Bearer sk-or-v1-abc123');
-  });
-
-  it('uses COPILOT_GITHUB_TOKEN (not COPILOT_API_KEY) for /models GET in BYOK+token mode', () => {
-    const adapter = createCopilotAdapter({
-      COPILOT_GITHUB_TOKEN: 'gho_oauth_token',
-      COPILOT_API_KEY: 'sk-or-v1-abc123',
-    });
-    const headers = adapter.getAuthHeaders(fakeModelsReq);
-    expect(headers['Authorization']).toBe('Bearer gho_oauth_token');
-  });
-
-  it('uses API key for /models GET when no GITHUB_TOKEN is set (BYOK-only mode)', () => {
-    const adapter = createCopilotAdapter({ COPILOT_API_KEY: 'sk-or-v1-abc123' });
-    const headers = adapter.getAuthHeaders(fakeModelsReq);
-    // In BYOK-only mode, githubToken is undefined so falls through to authToken (apiKey)
-    expect(headers['Authorization']).toBe('Bearer sk-or-v1-abc123');
-  });
-
-  it('is enabled when only COPILOT_API_KEY is set', () => {
-    const adapter = createCopilotAdapter({ COPILOT_API_KEY: 'sk-or-v1-abc123' });
-    expect(adapter.isEnabled()).toBe(true);
-  });
-
-  it('uses custom COPILOT_INTEGRATION_ID when set', () => {
-    const adapter = createCopilotAdapter({
-      COPILOT_API_KEY: 'sk-or-v1-abc123',
-      COPILOT_INTEGRATION_ID: 'my-custom-integration',
-    });
-    const headers = adapter.getAuthHeaders(fakeReq);
-    expect(headers['Copilot-Integration-Id']).toBe('my-custom-integration');
-  });
-
-  it('uses COPILOT_API_BASE_PATH when configured', () => {
-    const adapter = createCopilotAdapter({
-      COPILOT_API_KEY: 'sk-or-v1-abc123',
-      COPILOT_API_BASE_PATH: '/api/v1/',
-    });
-    expect(adapter.getBasePath()).toBe('/api/v1');
-  });
-
-  it('defaults to empty base path when COPILOT_API_BASE_PATH is not set', () => {
-    const adapter = createCopilotAdapter({ COPILOT_API_KEY: 'sk-or-v1-abc123' });
-    expect(adapter.getBasePath()).toBe('');
-  });
-});
-
-describe('resolveOpenCodeRoute', () => {
-  const OPENAI_TARGET = 'api.openai.com';
-  const ANTHROPIC_TARGET = 'api.anthropic.com';
-  const COPILOT_TARGET = 'api.githubcopilot.com';
-  const OPENAI_BASE = '/v1';
-  const ANTHROPIC_BASE = '';
-
-  it('should route to OpenAI when OPENAI_API_KEY is set (highest priority)', () => {
-    const route = resolveOpenCodeRoute(
-      'sk-openai-key', 'sk-anthropic-key', 'gho_copilot-token',
-      OPENAI_TARGET, ANTHROPIC_TARGET, COPILOT_TARGET,
-      OPENAI_BASE, ANTHROPIC_BASE
-    );
-    expect(route).not.toBeNull();
-    expect(route.target).toBe(OPENAI_TARGET);
-    expect(route.headers['Authorization']).toBe('Bearer sk-openai-key');
-    expect(route.basePath).toBe(OPENAI_BASE);
-    expect(route.needsAnthropicVersion).toBe(false);
-  });
-
-  it('should route to Anthropic when only ANTHROPIC_API_KEY is set', () => {
-    const route = resolveOpenCodeRoute(
-      undefined, 'sk-anthropic-key', undefined,
-      OPENAI_TARGET, ANTHROPIC_TARGET, COPILOT_TARGET,
-      OPENAI_BASE, ANTHROPIC_BASE
-    );
-    expect(route).not.toBeNull();
-    expect(route.target).toBe(ANTHROPIC_TARGET);
-    expect(route.headers['x-api-key']).toBe('sk-anthropic-key');
-    expect(route.basePath).toBe(ANTHROPIC_BASE);
-    expect(route.needsAnthropicVersion).toBe(true);
-  });
-
-  it('should prefer OpenAI over Anthropic when both are set', () => {
-    const route = resolveOpenCodeRoute(
-      'sk-openai-key', 'sk-anthropic-key', undefined,
-      OPENAI_TARGET, ANTHROPIC_TARGET, COPILOT_TARGET,
-      OPENAI_BASE, ANTHROPIC_BASE
-    );
-    expect(route).not.toBeNull();
-    expect(route.target).toBe(OPENAI_TARGET);
-    expect(route.headers['Authorization']).toBe('Bearer sk-openai-key');
-    expect(route.needsAnthropicVersion).toBe(false);
-  });
-
-  it('should route to Copilot when only copilotToken is set', () => {
-    const route = resolveOpenCodeRoute(
-      undefined, undefined, 'gho_copilot-token',
-      OPENAI_TARGET, ANTHROPIC_TARGET, COPILOT_TARGET,
-      OPENAI_BASE, ANTHROPIC_BASE
-    );
-    expect(route).not.toBeNull();
-    expect(route.target).toBe(COPILOT_TARGET);
-    expect(route.headers['Authorization']).toBe('Bearer gho_copilot-token');
-    expect(route.basePath).toBeUndefined();
-    expect(route.needsAnthropicVersion).toBe(false);
-  });
-
-  it('should prefer Anthropic over Copilot when both are set', () => {
-    const route = resolveOpenCodeRoute(
-      undefined, 'sk-anthropic-key', 'gho_copilot-token',
-      OPENAI_TARGET, ANTHROPIC_TARGET, COPILOT_TARGET,
-      OPENAI_BASE, ANTHROPIC_BASE
-    );
-    expect(route).not.toBeNull();
-    expect(route.target).toBe(ANTHROPIC_TARGET);
-    expect(route.headers['x-api-key']).toBe('sk-anthropic-key');
-    expect(route.needsAnthropicVersion).toBe(true);
-  });
-
-  it('should return null when no credentials are available', () => {
-    const route = resolveOpenCodeRoute(
-      undefined, undefined, undefined,
-      OPENAI_TARGET, ANTHROPIC_TARGET, COPILOT_TARGET,
-      OPENAI_BASE, ANTHROPIC_BASE
-    );
-    expect(route).toBeNull();
-  });
-
-  it('should not set Authorization header for Anthropic route', () => {
-    const route = resolveOpenCodeRoute(
-      undefined, 'sk-anthropic-key', undefined,
-      OPENAI_TARGET, ANTHROPIC_TARGET, COPILOT_TARGET,
-      OPENAI_BASE, ANTHROPIC_BASE
-    );
-    expect(route).not.toBeNull();
-    expect(route.headers['Authorization']).toBeUndefined();
-  });
-
-  it('should not set x-api-key header for OpenAI route', () => {
-    const route = resolveOpenCodeRoute(
-      'sk-openai-key', undefined, undefined,
-      OPENAI_TARGET, ANTHROPIC_TARGET, COPILOT_TARGET,
-      OPENAI_BASE, ANTHROPIC_BASE
-    );
-    expect(route).not.toBeNull();
-    expect(route.headers['x-api-key']).toBeUndefined();
-  });
-});
-
-// ── OpenCode adapter delegation ────────────────────────────────────────────────
-// Tests that verify OpenCode correctly delegates to its candidate adapters and
-// that all providers can be simultaneously active on their own ports.
-
-describe('OpenCode adapter delegation', () => {
-  const { createOpenCodeAdapter } = require('./providers/opencode');
-
-  function makeStubAdapter(name, enabled, { targetHost = `api.${name}.com`, basePath = '', authHeaders = {}, bodyTransform = null, urlTransform = undefined } = {}) {
-    return {
-      name,
-      isEnabled: () => enabled,
-      getTargetHost: () => targetHost,
-      getBasePath: () => basePath,
-      getAuthHeaders: () => authHeaders,
-      getBodyTransform: () => bodyTransform,
-      transformRequestUrl: urlTransform,
-    };
-  }
-
-  const fakeReq = { headers: {}, method: 'POST', url: '/v1/messages' };
-
-  it('routes to the first enabled candidate when multiple are configured', () => {
-    const openai     = makeStubAdapter('openai',    true,  { targetHost: 'api.openai.com',    basePath: '/v1', authHeaders: { Authorization: 'Bearer sk-oai' } });
-    const anthropic  = makeStubAdapter('anthropic', true,  { targetHost: 'api.anthropic.com', authHeaders: { 'x-api-key': 'sk-ant' } });
-    const copilot    = makeStubAdapter('copilot',   true,  { targetHost: 'api.githubcopilot.com', authHeaders: { Authorization: 'Bearer gho_cop' } });
-
-    const adapter = createOpenCodeAdapter({ AWF_ENABLE_OPENCODE: 'true' }, { candidateAdapters: [openai, anthropic, copilot] });
-
-    expect(adapter.isEnabled()).toBe(true);
-    expect(adapter.getTargetHost(fakeReq)).toBe('api.openai.com');
-    expect(adapter.getAuthHeaders(fakeReq).Authorization).toBe('Bearer sk-oai');
-    expect(adapter.getBasePath(fakeReq)).toBe('/v1');
-  });
-
-  it('skips disabled candidates and picks the next enabled one', () => {
-    const openai    = makeStubAdapter('openai',    false, { targetHost: 'api.openai.com' });
-    const anthropic = makeStubAdapter('anthropic', true,  { targetHost: 'api.anthropic.com', authHeaders: { 'x-api-key': 'sk-ant' } });
-
-    const adapter = createOpenCodeAdapter({ AWF_ENABLE_OPENCODE: 'true' }, { candidateAdapters: [openai, anthropic] });
-
-    expect(adapter.isEnabled()).toBe(true);
-    expect(adapter.getTargetHost(fakeReq)).toBe('api.anthropic.com');
-    expect(adapter.getAuthHeaders(fakeReq)['x-api-key']).toBe('sk-ant');
-  });
-
-  it('is disabled when all candidate adapters are disabled', () => {
-    const openai    = makeStubAdapter('openai',    false, {});
-    const anthropic = makeStubAdapter('anthropic', false, {});
-
-    const adapter = createOpenCodeAdapter({ AWF_ENABLE_OPENCODE: 'true' }, { candidateAdapters: [openai, anthropic] });
-    expect(adapter.isEnabled()).toBe(false);
-  });
-
-  it('is disabled when AWF_ENABLE_OPENCODE is not set, even if candidates are enabled', () => {
-    const openai = makeStubAdapter('openai', true, { targetHost: 'api.openai.com' });
-
-    const adapter = createOpenCodeAdapter({}, { candidateAdapters: [openai] });
-    expect(adapter.isEnabled()).toBe(false);
-  });
-
-  it('delegates body transform to the active candidate adapter', () => {
-    const transform = (buf) => Buffer.from(buf.toString().toUpperCase());
-    const anthropic = makeStubAdapter('anthropic', true, { targetHost: 'api.anthropic.com', bodyTransform: transform });
-
-    const adapter = createOpenCodeAdapter({ AWF_ENABLE_OPENCODE: 'true' }, { candidateAdapters: [anthropic] });
-    const fn = adapter.getBodyTransform();
-    expect(fn).toBe(transform);
-  });
-
-  it('returns null body transform when active candidate has none', () => {
-    const openai = makeStubAdapter('openai', true, { bodyTransform: null });
-    const adapter = createOpenCodeAdapter({ AWF_ENABLE_OPENCODE: 'true' }, { candidateAdapters: [openai] });
-    expect(adapter.getBodyTransform()).toBeNull();
-  });
-
-  it('delegates URL transform to the active candidate when one is defined', () => {
-    const urlTransform = (url) => url.replace('?key=placeholder', '');
-    const gemini = makeStubAdapter('gemini', true, { targetHost: 'generativelanguage.googleapis.com', urlTransform });
-
-    const adapter = createOpenCodeAdapter({ AWF_ENABLE_OPENCODE: 'true' }, { candidateAdapters: [gemini] });
-    const transformed = adapter.transformRequestUrl('/v1/models?key=placeholder');
-    expect(transformed).toBe('/v1/models');
-  });
-
-  it('returns url unchanged when active candidate has no URL transform', () => {
-    const openai = makeStubAdapter('openai', true, {});
-    const adapter = createOpenCodeAdapter({ AWF_ENABLE_OPENCODE: 'true' }, { candidateAdapters: [openai] });
-    expect(adapter.transformRequestUrl('/v1/chat/completions')).toBe('/v1/chat/completions');
-  });
-
-  it('reports the active adapter name at startup for introspection', () => {
-    const anthropic = makeStubAdapter('anthropic', false, {});
-    const copilot   = makeStubAdapter('copilot',   true,  { targetHost: 'api.githubcopilot.com' });
-
-    const adapter = createOpenCodeAdapter({ AWF_ENABLE_OPENCODE: 'true' }, { candidateAdapters: [anthropic, copilot] });
-    expect(adapter._startupActiveAdapterName).toBe('copilot');
-  });
-
-  it('exposes the candidate adapter list for introspection', () => {
-    const openai    = makeStubAdapter('openai',    true, {});
-    const anthropic = makeStubAdapter('anthropic', true, {});
-
-    const adapter = createOpenCodeAdapter({ AWF_ENABLE_OPENCODE: 'true' }, { candidateAdapters: [openai, anthropic] });
-    expect(adapter._candidateAdapters).toHaveLength(2);
-    expect(adapter._candidateAdapters[0].name).toBe('openai');
-    expect(adapter._candidateAdapters[1].name).toBe('anthropic');
-  });
-
-  it('all providers remain independently active on their own ports', () => {
-    // Simulate the production setup: OpenAI + Anthropic + Copilot all enabled
-    const openai    = makeStubAdapter('openai',    true, { targetHost: 'api.openai.com' });
-    const anthropic = makeStubAdapter('anthropic', true, { targetHost: 'api.anthropic.com' });
-    const copilot   = makeStubAdapter('copilot',   true, { targetHost: 'api.githubcopilot.com' });
-
-    const opencode = createOpenCodeAdapter({ AWF_ENABLE_OPENCODE: 'true' }, { candidateAdapters: [openai, anthropic, copilot] });
-
-    // Each provider is independently enabled
-    expect(openai.isEnabled()).toBe(true);
-    expect(anthropic.isEnabled()).toBe(true);
-    expect(copilot.isEnabled()).toBe(true);
-
-    // OpenCode routes to the first enabled (OpenAI in this priority order)
-    expect(opencode.isEnabled()).toBe(true);
-    expect(opencode.getTargetHost(fakeReq)).toBe('api.openai.com');
-
-    // All three base providers are still individually reachable (different ports)
-    expect(openai.getTargetHost()).toBe('api.openai.com');
-    expect(anthropic.getTargetHost()).toBe('api.anthropic.com');
-    expect(copilot.getTargetHost()).toBe('api.githubcopilot.com');
-  });
-});
+const { composeBodyTransforms } = require('./proxy-utils');
+const { createCopilotAdapter } = require('./providers/copilot');
 
 describe('httpProbe', () => {
   let server;
@@ -2784,112 +1505,6 @@ describe('provider adapter alwaysBind', () => {
     const { statusCode, body } = adapter.getUnconfiguredHealthResponse();
     expect(statusCode).toBe(503);
     expect(body.error).toMatch(/no candidate provider credentials/);
-  });
-});
-
-// ── proxyRequest X-Initiator billing header injection ─────────────────────────
-//
-// When forwarding to the Copilot API, the proxy must inject "X-Initiator: agent"
-// when the header is absent so that autonomous turns are billed correctly instead
-// of defaulting to the (more expensive) "user" rate.
-//
-describe('proxyRequest X-Initiator injection', () => {
-  /** Minimal mock for http.IncomingMessage backed by EventEmitter. */
-  function makeReq(headers = {}) {
-    const req = new EventEmitter();
-    req.url = '/v1/chat/completions';
-    req.method = 'POST';
-    req.headers = { 'content-type': 'application/json', ...headers };
-    return req;
-  }
-
-  /** Minimal mock for http.ServerResponse. */
-  function makeRes() {
-    return {
-      headersSent: false,
-      setHeader: jest.fn(),
-      writeHead: jest.fn(),
-      end: jest.fn(),
-    };
-  }
-
-  afterEach(() => {
-    jest.restoreAllMocks();
-  });
-
-  /**
-   * Mock https.request to capture the outgoing options (including headers)
-   * without making a real network connection.
-   */
-  function mockHttpsRequest() {
-    let capturedOptions;
-    jest.spyOn(https, 'request').mockImplementation((options) => {
-      capturedOptions = options;
-      const proxyReq = new EventEmitter();
-      proxyReq.end = jest.fn();
-      proxyReq.write = jest.fn();
-      proxyReq.destroy = jest.fn();
-      return proxyReq;
-    });
-    return { getCaptured: () => capturedOptions };
-  }
-
-  it('injects x-initiator: agent when absent on direct copilot requests', () => {
-    const { getCaptured } = mockHttpsRequest();
-    const req = makeReq();
-    proxyRequest(req, makeRes(), 'api.githubcopilot.com', { 'Authorization': 'Bearer token' }, 'copilot');
-    req.emit('end');
-    expect(getCaptured().headers['x-initiator']).toBe('agent');
-  });
-
-  it('injects x-initiator: agent when absent on enterprise githubcopilot.com target', () => {
-    const { getCaptured } = mockHttpsRequest();
-    const req = makeReq();
-    proxyRequest(req, makeRes(), 'api.enterprise.githubcopilot.com', { 'Authorization': 'Bearer token' }, 'copilot');
-    req.emit('end');
-    expect(getCaptured().headers['x-initiator']).toBe('agent');
-  });
-
-  it('injects x-initiator: agent when OpenCode routes to Copilot backend', () => {
-    // OpenCode delegates to Copilot; provider name is "opencode" but target host is Copilot.
-    const { getCaptured } = mockHttpsRequest();
-    const req = makeReq();
-    proxyRequest(req, makeRes(), 'api.githubcopilot.com', { 'Authorization': 'Bearer token' }, 'opencode');
-    req.emit('end');
-    expect(getCaptured().headers['x-initiator']).toBe('agent');
-  });
-
-  it('preserves a client-supplied x-initiator value on copilot requests', () => {
-    const { getCaptured } = mockHttpsRequest();
-    const req = makeReq({ 'x-initiator': 'user' });
-    proxyRequest(req, makeRes(), 'api.githubcopilot.com', { 'Authorization': 'Bearer token' }, 'copilot');
-    req.emit('end');
-    expect(getCaptured().headers['x-initiator']).toBe('user');
-  });
-
-  it('preserves a client-supplied x-initiator value on OpenCode→Copilot requests', () => {
-    const { getCaptured } = mockHttpsRequest();
-    const req = makeReq({ 'x-initiator': 'user' });
-    proxyRequest(req, makeRes(), 'api.githubcopilot.com', { 'Authorization': 'Bearer token' }, 'opencode');
-    req.emit('end');
-    expect(getCaptured().headers['x-initiator']).toBe('user');
-  });
-
-  it('does not inject x-initiator for non-copilot provider targets', () => {
-    const { getCaptured } = mockHttpsRequest();
-    const req = makeReq();
-    proxyRequest(req, makeRes(), 'api.anthropic.com', { 'x-api-key': 'sk-ant-test' }, 'anthropic');
-    req.emit('end');
-    expect(getCaptured().headers['x-initiator']).toBeUndefined();
-  });
-
-  it('does not inject x-initiator when OpenCode routes to non-Copilot backend', () => {
-    // OpenCode routing to OpenAI — no Copilot billing header should be injected.
-    const { getCaptured } = mockHttpsRequest();
-    const req = makeReq();
-    proxyRequest(req, makeRes(), 'api.openai.com', { 'Authorization': 'Bearer sk-test' }, 'opencode');
-    req.emit('end');
-    expect(getCaptured().headers['x-initiator']).toBeUndefined();
   });
 });
